+++ js/background/gmail/fetch.js
@@ -15,8 +15,7 @@ this.retrieveEmailThread = function (threadId) {
     })
         .then(pluckEmailsFromThread)
         .then(filterOutSentEmails)
-        .then(processEmails)
-        .then(fetchSafeBrowsingInformation);
+        .then(processEmails);
 };
 
 /**